refactor(KakaoLogin, Group): 	1. 카카오 로그인 - 카카오서버로부터 id값 받아와 Long 타입 memberId로 삽입 	2. 기존 Integer 타입의 memberId 사용하던 곳 Long 타입으로 수정 	3. 그룹 초대 및 승인 코드 수정 완료
issue: #3

+++ backend/src/main/java/com/patriot/fourlipsclover/group/repository/GroupJoinRequestRepository.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface GroupJoinRequestRepository extends JpaRepository<GroupJoinRequest, Long> {
+public interface GroupJoinRequestRepository extends JpaRepository<GroupJoinRequest, Integer> {
 
-    Optional<GroupJoinRequest> findByGroup_GroupIdAndMember_MemberIdAndToken(Integer groupId, Integer memberId, String token);
+    Optional<GroupJoinRequest> findByGroup_GroupIdAndMember_MemberIdAndToken(Integer groupId, Long memberId, String token);
 
 }
